Global Constants - Implemenation
What it does?
Introduces a new Global Constant implementation that enables to replace direct hardcoded strings in HTML, TS to be fetched from Constants.

How does it affect?
It enables easier access of constants for reading, editing and deleting.
Things will be done in one place and hence can be reused across.

Changes in this commit:

1. A generic service constants.service.ts that handles the rendering of messages of the desired component
2. A constants-mapper.ts that maps the constants file of a component against a name that can be used for retrieving.
3. n constant files each representing a component.
4. A global-constants.ts file that includes all the constants files.

Tracks #33.